requirements.txt
absl-py==1.2.0
accelerate==0.12.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==0.4.3
altair==4.2.0
analytics-python==1.4.0
antlr4-python3-runtime==4.8
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
backoff==1.10.0
basicsr==1.4.2
bcrypt==4.0.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
clean-fid==0.1.30
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
coloredlogs==15.0.1
commonmark==0.9.1
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
diffusers==0.2.4
docker-pycreds==0.4.0
docopt==0.6.2
e==1.4.5
einops==0.3.0
entrypoints==0.4
facexlib==0.2.5
fastapi==0.82.0
ffmpy==0.3.0
filelock==3.8.0
filterpy==1.4.5
Flask==2.2.2
flatbuffers==2.0.7
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
-e git+https://github.com/TencentARC/GFPGAN@8d2447a2d918f8eba5a4a01463fd48e45126a379#egg=gfpgan
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
gradio==3.2
grpcio==1.48.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.9.1
humanfriendly==10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==4.12.0
invisible-watermark==0.1.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonable==0.3.1
jsonmerge==1.8.0
jsonschema==4.14.0
-e git+https://github.com/hlky/k-diffusion-sd@b435d7a7818672cd616a282aeb018fc4bdb4b778#egg=k_diffusion
kdtree==0.16
kiwisolver==1.4.4
kornia==0.6.0
-e git+ssh://git@github.com/cloudwow/stable-diffusion.git@69ae4b35e0a0f6ee1af8bb9a5d0016ccb27e36dc#egg=latent_diffusion
linkify-it-py==1.0.3
llvmlite==0.39.1
lmdb==1.3.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mdit-py-plugins==0.3.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.2.1
multidict==6.0.2
mwcli==0.0.3
mwtypes==0.3.2
mwxml==0.3.3
networkx==2.8.6
numba==0.56.2
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
orjson==3.8.0
packaging==21.3
pandas==1.4.4
para==0.0.8
paramiko==2.11.0
pathtools==0.1.2
Peacasso==0.0.11a0
Pillow==9.2.0
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pudb==2019.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydeck==0.8.0b1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.13.0
pymongo==4.2.0
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
-e git+https://github.com/xinntao/Real-ESRGAN@e5763af5749430c9f7389f185cc53f90c4852ed5#egg=realesrgan
regex==2022.8.17
requests==2.25.1
requests-oauthlib==1.3.1
resize-right==0.0.2
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
semver==2.13.0
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.19.1
streamlit==1.12.2
sympy==1.11.1
-e git+https://github.com/CompVis/taming-transformers.git@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tb-nightly==2.11.0a20220907
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2022.8.12
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1+cu113
torch-fidelity==0.3.0
torchaudio==0.12.1+cu113
torchdiffeq==0.2.3
torchmetrics==0.6.0
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.0
transformers==4.19.2
typer==0.6.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uc-micro-py==1.0.1
urllib3==1.26.12
urwid==2.1.2
uvicorn==0.18.3
validators==0.20.0
wandb==0.13.2
watchdog==2.1.9
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.2.2
wikitextparser==0.50.1
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1
